Rename storage file to dynamodb
